Config some data

+++ app/src/main/java/com/example/ihomebyme/db/entity/FundEntity.kt
@@ -4,7 +4,7 @@ import androidx.room.Entity
 import androidx.room.PrimaryKey
 
 @Entity(tableName = "funds")
-data class FundEntity(var fundName: String, var fundPay: Double) {
+data class FundEntity(var fundName: String, var fundPay: Long) {
 
     @PrimaryKey(autoGenerate = true)
     var id: Long? = null
